Change Amount type To Long

+++ InternetBank.Core/DTO/RegisterAccountDTO.cs
@@ -7,5 +7,5 @@ public class RegisterAccountDTO
     [Required]
     public AccountType accountType { get; set; }
     [Required, Range(100000, int.MaxValue, ErrorMessage = "حداقل مقدار برای افتتاح حساب 100000 ریال می باشد")]
-    public int Amount { get; set; }
+    public long Amount { get; set; }
 }
